Move createCourtCase from JudgeOverview to useCase hook
